Improve the upload report output
1) drop already released uploads from proposed

This makes the list more readable by eliminating redundancy.
And it makes the list in proposed really the one we want/need help on
testing as the others are already out.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>

2) improve headers for Proposed and SRU

In proposed we want to call for support and in SRU we want to make it
more clear that this completed already.

Fixes 10

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>

3) Link to uploads of uploads

By that people can easily get to the changed content as well as from
there to the full bug descriptions related to the upload.

Fixes 9

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>